added code to parse code and refactored to utils.p
