+++ parsers/parse_description.py
@@ -1,72 +1,6 @@
-import pandas as pd
-import sklearn as sk
-import math 
-import nltk
-from collections import Counter
-import os.path
-from gensim import corpora
 from gensim.models import LsiModel
-from nltk.tokenize import RegexpTokenizer
-from nltk.stem.porter import PorterStemmer
 from gensim.models.coherencemodel import CoherenceModel
-import matplotlib.pyplot as plt
-import requests
-
-stopwords_list = requests.get("https://gist.githubusercontent.com/rg089/35e00abf8941d72d419224cfd5b5925d/raw/12d899b70156fd0041fa9778d657330b024b959c/stopwords.txt").content
-stop_words = set(stopwords_list.decode().splitlines()) 
-
-def split_text(text:str) -> list:
-    return text.split(" ")
-
-def remove_stopwords(wordList: list) -> list:
-    filtered_sentence = [w for w in wordList if not w in stop_words]
-    return filtered_sentence
-
-def generate_tf_idf(text: str) -> Counter:
-    formatted_text = remove_stopwords(split_text(text))
-    count = len(formatted_text)
-    formatted_text_dict = Counter(formatted_text)
-    for word in formatted_text_dict:
-        formatted_text_dict[word] = formatted_text_dict[word]/count
-    return formatted_text_dict
-
-def preprocess_data(doc_set):
-    """
-    Input  : docuemnt list
-    Purpose: preprocess text (tokenize, removing stopwords, and stemming)
-    Output : preprocessed text
-    """
-    # initialize regex tokenizer
-    tokenizer = RegexpTokenizer(r'\w+')
-    # Create p_stemmer of class PorterStemmer
-    p_stemmer = PorterStemmer()
-    # list for tokenized documents in loop
-    texts = []
-    #removes stop words from document list
-    doc_set = remove_stopwords(doc_set)
-    # loop through document list
-    for i in doc_set:
-        # clean and tokenize document string
-        raw = i.lower()
-        tokens = tokenizer.tokenize(raw)
-        # stem tokens
-        stemmed_tokens = [p_stemmer.stem(i) for i in tokens]
-        # add tokens to list
-        texts.append(stemmed_tokens)
-    return texts
-
-def prepare_corpus(doc_clean):
-    """
-    Input  : clean document
-    Purpose: create term dictionary of our courpus and Converting list of documents (corpus) into Document Term Matrix
-    Output : term dictionary and Document Term Matrix
-    """
-    # Creating the term dictionary of our courpus, where every unique term is assigned an index. dictionary = corpora.Dictionary(doc_clean)
-    dictionary = corpora.Dictionary(doc_clean)
-    # Converting list of documents (corpus) into Document Term Matrix using dictionary prepared above.
-    doc_term_matrix = [dictionary.doc2bow(doc) for doc in doc_clean]
-    # generate LDA model
-    return dictionary,doc_term_matrix
+from utils import prepare_corpus
 
 def create_gensim_lsa_model(doc_clean,number_of_topics):
     """
